Tidied, Variable length arrays new/deleted, made casting clear

+++ src/SMC_class.cpp
@@ -5,7 +5,7 @@
    you may not use this file except in compliance with the License.
    You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+	http://www.apache.org/licenses/LICENSE-2.0
 
    Unless required by applicable law or agreed to in writing, software
    distributed under the License is distributed on an "AS IS" BASIS,
@@ -22,7 +22,6 @@
 
    Jonathan Petticrew, University of Sheffield, 2017.
  */
-
 #include "SMC.h"
 #include <math.h>
 #include <stdio.h>
@@ -161,6 +160,6 @@ void SMC::mat(int x){
 	}
 
 	Emax=MAX_eV*q;
-	NUMPOINTS = (MAX_eV*1000);
+	NUMPOINTS = (int)(MAX_eV*1000);
 	N=(1/(exp(hw/(K*T))-1));
 };
